+++ eccodes/eccodes.py
@@ -61,6 +61,7 @@ from gribapi import (
 )
 from gribapi import grib_clone as codes_clone
 from gribapi import grib_context_delete as codes_context_delete
+from gribapi import grib_context_set_logging as codes_context_set_logging
 from gribapi import grib_copy_namespace as codes_copy_namespace
 from gribapi import grib_count_in_file as codes_count_in_file
 from gribapi import grib_find_nearest as codes_grib_find_nearest
@@ -371,6 +372,7 @@ __all__ = [
     "codes_skip_read_only",
     "codes_write",
     "codes_context_delete",
+    "codes_context_set_logging",
     "CodesInternalError",
     "ConceptNoMatchError",
     "ConstantFieldError",