+++ gribapi/grib_api.h
@@ -139,6 +139,7 @@ void grib_multi_support_on(grib_context* c);
 void grib_multi_support_off(grib_context* c);
 void grib_multi_support_reset_file(grib_context* c, FILE* f);
 long grib_get_api_version(void);
+void grib_context_set_logging_file(grib_context* c, FILE* f);
 
 char* grib_samples_path(const grib_context *c);
 char* grib_definition_path(const grib_context *c);